+++ android/app/src/main/kotlin/coop/polypoly/polypod/AppUpdateReceiver.kt
@@ -5,9 +5,8 @@ import android.content.Context
 import android.content.Intent
 import androidx.work.OneTimeWorkRequest
 import androidx.work.WorkManager
+import coop.polypoly.polypod.core.UpdateNotification
 import coop.polypoly.polypod.logging.LoggerFactory
-import coop.polypoly.polypod.updatenotification.PushNotificationWorker
-import coop.polypoly.polypod.updatenotification.UpdateNotification
 import java.util.concurrent.TimeUnit
 
 class AppUpdateReceiver : BroadcastReceiver() {
