+++ android/app/src/main/kotlin/coop/polypoly/polypod/MainActivity.kt
@@ -4,8 +4,8 @@ import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import coop.polypoly.polypod.core.UpdateNotification
 import coop.polypoly.polypod.features.FeatureStorage
-import coop.polypoly.polypod.updatenotification.UpdateNotification
 
 class MainActivity : AppCompatActivity() {
