+++ android/app/src/test/kotlin/coop/polypoly/polypod/updatenotification/UpdateNotificationTest.kt
@@ -3,6 +3,7 @@ package coop.polypoly.polypod.updatenotification
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
+import coop.polypoly.polypod.core.UpdateNotification
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
